Methods for conversions between cytoframe/cytoset and flowFrame/flowSet

These methods consist of a pair of methods to coerce a cytoframe to or from a flowFrame and another pair to coerce a cytoset to or from a flowSet.

cytoframe_to_flowFrame(fr)

flowFrame_to_cytoframe(fr, ...)

cytoset_to_flowSet(cs)

flowSet_to_cytoset(fs, path = tempfile())

Details

The conversion between the two sets of data container classes mostly entails a conversion of the back-end representation of the data. cytoframe and cytoset objects contain flowFrame and flowSet objects respectively, so coercion of a cytoframe to flowFrame entails moving the data from the 'C'-level data structure to the corresponding exprs, description, and parameters slots. Coercion of a flowFrame to a cytoframe entails creation of the 'C'-level data structure from the flowFrame slots. The names of each of the methods are pretty self-explanatory.

Methods

cytoframe_to_flowFrame(object = "cytoframe")

Returns a flowFrame object coerced from a cytoframe object.

flowFrame_to_cytoframe(object = "flowFrame")

Returns a cytoframe object coerced from a flowFrame object.

cytoset_to_flowSet(object = "cytoset")

Returns a flowSet object coerced from a cytoset object.

flowSet_to_cytoset(object = "flowSet")

Returns a cytoset object coerced from a flowSet object.

Examples

library(flowCore)
data("GvHD")
fs <- GvHD[1]
cs <- flowSet_to_cytoset(fs)
cf <- cs[[1, returnType="cytoframe"]]
ff <- cytoframe_to_flowFrame(cf)