Methods to change channel and marker names for cytoframe and cytoset objects

The methods allow direct alteration of channel names or marker names of cytoframe and cytoset objects. These objects are accessed by reference and changed in place, so there is no need to assign the return value of these methods.

cf_swap_colnames(x, col1, col2)

cf_rename_channel(x, old, new)

cf_rename_marker(x, old, new)

cs_swap_colnames(x, col1, col2)

Arguments

x
a cytoframe

col1
first channel name to swap

col2
second channel name to swap

old
old channel or marker name to be changed

new
new channel or marker name after change