Get the names of all nodes from a gating hierarchy.

gs_get_pop_paths returns a character vector of names of the nodes (populations) in the GatingSet.

gs_get_pop_paths(x, y = NULL, order = "regular", path = "full",
  showHidden = FALSE, ...)

gh_get_pop_paths(x, y = NULL, order = "regular", path = "full",
  showHidden = FALSE, ...)

Arguments

x
A GatingSet Assuming the gating hierarchy are identical within the GatingSet, the Gating tree of the first sample is used to query the node information.

y
A character not used.

order
order=c("regular","tsort","bfs") returns the nodes in regular, topological or breadth-first sort order. "regular" is default.

path
A character or numeric scalar. when numeric, it specifies the fixed length of gating path (length 1 displays terminal name). When character, it can be either 'full' (full path, which is default) or 'auto' (display the shortest unique gating path from the bottom of gating tree).

showHidden
logical whether to include the hidden nodes

...
Additional arguments.

Value

gs_get_pop_paths returns a character vector of node/population names, ordered appropriately.

Details

integer indices of nodes are based on regular order,so whenver need to map from character node name to integer node ID,make sure to use default order which is regular.

Examples