Return the flowCore gate definition associated with a node in a GatingHierarchy/GatingSet.

Return the flowCore gate definition object associated with a node in a GatingHierarchy or GatingSet object.

gh_pop_get_gate(obj, y)

gs_pop_get_gate(obj, y)

Arguments

obj
A GatingHierrarchy or GatingSet

y
A character the name or full(/partial) gating path of the node of interest.

Value

A gate object from flowCore. Usually a polygonGate, but may be a rectangleGate. Boolean gates are represented by a "BooleanGate" S3 class. This is a list boolean gate definition that references populations in the GatingHierarchy and how they are to be combined logically. If obj is a GatingSet, assuming the trees associated with each GatingHierarchy are identical, then this method will return a list of gates, one for each sample in the GatingSet corresponding to the same population indexed by y.

See also

Examples

  if (FALSE)   #gh is a GatingHierarchy
    gh_pop_get_gate(gh, "CD3") #return the gate for the fifth node in the tree, but fetch it by name.
    #G is a GatingSet
    gs_pop_get_gate(G, "CD3") #return a list of gates for the fifth node in each tree

#> Error in h(simpleError(msg, call)): error in evaluating the argument 'X' in selecting a method for function 'lapply': object 'G' not found