Gating-ML 2.0 Log transformation.

Used to construct GML 2.0 flog transformer object.

logtGml2_trans(t = 262144, m = 4.5, n = 6, equal.space = FALSE)

Arguments

t
top scale value

m
number of decades

n
desired number of breaks (the actual number will be different depending on the data range)

equal.space
whether breaks at equal-spaced intervals

Value

logtGml2 transformation object

Details

GML 2.0 standard log transform function constructor. The definition is as in the GML 2.0 standard section 6.2 "parametrized logarithmic transformation -- flog" This deviates from standard only in the following way. Before applying the logarithmic transformation, non-positive values are assigned the smallest positive value from the input rather than having undefined values (NA) under the transformation.

Examples

trans.obj <- logtGml2_trans(t = 1e3, m = 1, equal.space = TRUE)
data <- 1:1e3
brks.func <- trans.obj[["breaks"]]
brks <- brks.func(data)
brks # fasinh space displayed at raw data scale

#> [1]    1.000000    3.981072   15.848932   63.095734  251.188643 1000.000000

#transform it to verify it is equal-spaced at transformed scale
trans.func <- trans.obj[["transform"]]
brks.trans <- trans.func(brks)
brks.trans

#> [1] -2.0 -1.4 -0.8 -0.2  0.4  1.0