Add populations to a GatingHierarchy

Add populations to a GatingHierarchy

pop_add(gate, gh, ...)

# S3 method for filter
pop_add(gate, gh, ...)

# S3 method for filters
pop_add(gate, gh, names = NULL, ...)

# S3 method for quadGate
pop_add(gate, gh, names = NULL, ...)

# S3 method for logical
pop_add(gate, gh, parent, name, recompute, cluster_method_name = NULL, ...)

# S3 method for factor
pop_add(gate, gh, name = NULL, ...)

# S3 method for logicalFilterResult
pop_add(gate, gh, ...)

# S3 method for multipleFilterResult
pop_add(gate, gh, name = NULL, ...)

gh_pop_remove(gh, node, ...)

Arguments

gate
a gate object that extends flowCore::filter or flowCore::filters

gh
GatingHierarchy

...
other arguments

names
a character vector of length four,which specifies the population names resulted by adding a quadGate.The order of the names is clock-wise starting from the top left quadrant population.

parent
a character scalar to specify the parent node name where the new gate to be added to, by default it is NULL,which indicates the root node

name
the population name

recompute
whether to recompute the gates

cluster_method_name
when adding the logical vectors as the gates, the name of the cluster method can be used to tag the populations as the extra meta information associated with the gates.

node
population name/path