Compute the cell events by the gates stored within the gating tree.

Compute each cell event to see if it falls into the gate stored within the gating tree and store the result as cell count.

recompute(x, y="root", alwaysLoadData=FALSE, ...)

Arguments

x
GatingSet or GatingSetList

y
character node name or node path. Default "root". Optional.

alwaysLoadData
logical. Specifies whether to load the flow raw data for gating boolean gates. Default 'FALSE'. Optional. Sometime it is more efficient to skip loading the raw data if all the reference nodes and parent are already gated. 'FALSE' will check the parent node and reference to determine whether to load the data. This check may not be sufficient since the further upstream ancestor nodes may not be gated yet. In that case, we allow the gating to fail and prompt user to recompute those nodes explictily. When TRUE, then it forces data to be loaded to guarantee the gating process to be uninterrupted at the cost of unnecessary data IO.

...
other arguments leaf.bool whether to compute the leaf boolean gate, default is TRUE

Details

It is usually used immediately after add or gs_pop_set_gate calls.