save/load a cytoset to/from disk. load_cytoset() can load a cytoset from either the archive previously saved by save_cytoset() call or from a folder that contains a collection of inidivudal cytoframe files (either in h5 format or tiledb format)

Save/load a cytoset to/from the disk.

save_cytoset(cs, path, ...)

load_cytoset(path, verbose = FALSE, ...)

Arguments

cs
A cytoset

path
A character scalar giving the path to save/load the cytoset to/from.

...
other arguments passed to save_gs/load_gs

verbose
whether to print details. Default is FALSE.

Value

load_cytoset returns a cytoset object

Examples

if (FALSE) {
  #cs is a cytoset
  save_cytoset(cs, outdir)
  cs <-load_cytoset(outdir)

#or from cytoframe on-disk files
# e.g. h5_dir contains the cytoframes in h5 format
cs <- load_cytoset(h5_dir)

}