save/load a GatingSet/GatingSetList to/from disk.

Save/load a GatingSet/GatingSetList which is the gated flow data including gates and populations to/from the disk. The GatingSet object The internal C data structure (gating tree),ncdfFlowSet object(if applicable)

save_gs(gs, path, cdf = c("copy", "move", "skip", "symlink", "link"), ...)

load_gs(path, h5_readonly = TRUE, select = character(), verbose = FALSE)

save_gslist(gslist, path, ...)

load_gslist(path)

Arguments

path
A character scalar giving the path to save/load the GatingSet to/from.

cdf
a character scalar. The valid options are :"copy","move","skip","symlink" specifying what to do with the cdf data file. Sometimes it is more efficient to move or create a symlink of the existing cdf file to the archived folder. It is useful to "skip" archiving cdf file if raw data has not been changed.

...
other arguments: not used.

h5_readonly
whether to open h5 data as read-only. Default is TRUE

select
an integer or character vector to select a subset of samples to load

verbose
logical flag to optionally print the versions of the libraries that were used to archive the GatingSet for troubleshooting purpose.

gslist
A GatingSetList

G
A GatingSet

Value

load_gs returns a GatingSet object load_gslist returns a GatingSetList object

See also

Examples

if (FALSE) {
        #G is a GatingSet
        save_gs(G,path="tempFolder")
        G1<-load_gs(path="tempFolder")

        #G is a GatingSet

        save_gslist(gslist1,path="tempFolder")
        gslist2<-load_gslist(path="tempFolder")
}