Constructor for transformerList object

Similar to transformList function, it constructs a list of transformer objects generated by trans_new method from scales so that the inverse and breaks functions are also included.

transformerList(from, trans)

Arguments

from
channel names

trans
a trans object or a list of trans objects constructed by trans_new method.

Examples

library(flowCore)
library(scales)
#create tranformer object from scratch
trans <- logicleTransform(w = 0.5, t = 262144, m = 4.5, a = 0)
inv <- inverseLogicleTransform(trans = trans)
trans.obj <- flow_trans("logicle", trans, inv, n = 5, equal.space = FALSE)

#or simply use convenient constructor
#trans.obj <- logicle_trans(n = 5, equal.space = FALSE, w = 0.5, t = 262144, m = 4.5, a = 0)

transformerList(c("FL1-H", "FL2-H"), trans.obj)

#> $`FL1-H`
#> Transformer:  logicle 
#> 
#> $`FL2-H`
#> Transformer:  logicle 
#> 
#> attr(,"class")
#> [1] "transformerList" "list"           

#use different transformer for each channel
trans.obj2 <- asinhtGml2_trans()
transformerList(c("FL1-H", "FL2-H"), list(trans.obj, trans.obj2))

#> $`FL1-H`
#> Transformer:  logicle 
#> 
#> $`FL2-H`
#> Transformer:  asinhtGml2 
#> 
#> attr(,"class")
#> [1] "transformerList" "list"